Modified Global.asax to include SetupRefreshJob() which should fix the IIS7 dying out
Modified Global.asax to include SetupRefreshJob() which should fix the
IIS7 dying out in Godaddy shared hosting accounts. Because in hosting
accounts that are shared, we cannot set the IIS to always running but
instead is probably using a "On Demand" or "On Request" mode which might
be causing the initial slowness. This code fixes that. I still might
have to add a call to this method from BlogEngineConfiguration.cs in
App_Start folder. I will monitor it to see if it works.